requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
httpx==0.27.2
idna==3.6
pymongo==4.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
groq
langchain==0.1.16
google-genai
langchain-core
langchain-groq
langchain-openai
langchain-google-genai
langdetect
